build: added clang-c2 support and optimized the build process
X-Committed-on-Branch: topic/clangc2

X-Committed-on-Branch: (no branch, rebasing topic/clangc2)

+++ src/common/fileglobber.cc
@@ -30,7 +30,7 @@ namespace common {
 struct GlobPrivateData
 {
     bool nowildcard;
-    long handle;
+    intptr_t handle;
     struct _finddata_t fdata;
     char dir[_MAX_PATH];
     char tmpfname[_MAX_PATH];
